% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltb.R
\name{mycltb}
\alias{mycltb}
\title{CLT Binomial}
\usage{
mycltb(n, iter, p = 0.5, ...)
}
\arguments{
\item{n}{Sample size for each iteration.}

\item{iter}{Number of iterations.}

\item{p}{The probability of success for each trial in the Binomial distribution (default is 0.5).}

\item{...}{Additional arguments to be passed to plotting functions.}
}
\value{
A histogram of sample means and a theoretical normal distribution curve.
}
\description{
The Central Limit Theorem (CLT) applied to a Binomial distribution. This function generates a histogram of sample means obtained by sampling from a Binomial distribution, along with the corresponding theoretical normal distribution curve.
}
\examples{
# Example 1: CLT with default parameters
mycltb(n = 10, iter = 10000)

# Example 2: CLT with custom parameters
mycltb(n = 20, iter = 5000, p = 0.3, col = "blue")

}
