% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{CLT poisson.test
The Central Limit Theorem (CLT) applied to a Poisson distribution. This function generates a histogram of sample means obtained by sampling from a Poisson distribution, along with the corresponding theoretical normal distribution curve and the probability function for the Poisson distribution.}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{Sample size for each iteration.}

\item{iter}{Number of iterations.}

\item{lambda}{The mean parameter of the Poisson distribution (default is 10).}

\item{...}{Additional arguments to be passed to plotting functions.}
}
\value{
A histogram of sample means, a theoretical normal distribution curve, and a probability function for the Poisson distribution.
}
\description{
CLT poisson.test
The Central Limit Theorem (CLT) applied to a Poisson distribution. This function generates a histogram of sample means obtained by sampling from a Poisson distribution, along with the corresponding theoretical normal distribution curve and the probability function for the Poisson distribution.
}
\examples{
# Example 1: CLT with default parameters
mycltp(n = 10, iter = 10000)

# Example 2: CLT with custom parameters
mycltp(n = 20, iter = 5000, lambda = 15, col = "blue")


}
