% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltu.R
\name{mycltu}
\alias{mycltu}
\title{CLT Uniform}
\usage{
mycltu(n, iter, a = 0, b = 10)
}
\arguments{
\item{n}{Sample size for each iteration.}

\item{iter}{Number of iterations.}

\item{a}{Lower limit of the uniform distribution (default is 0).}

\item{b}{Upper limit of the uniform distribution (default is 10).}
}
\value{
A histogram of sample means, a theoretical normal distribution curve, and the probability density function of the Uniform distribution.
}
\description{
The Central Limit Theorem (CLT) applied to a Uniform distribution. This function generates a histogram of sample means obtained by sampling from a Uniform distribution, along with the corresponding theoretical normal distribution curve and the probability density function of the Uniform distribution.
}
\examples{
# Example 1: CLT with default parameters
mycltu(n = 10, iter = 10000)

# Example 2: CLT with custom parameters
mycltu(n = 20, iter = 5000, a = 5, b = 15)

}
