% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Area of Curve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{The mean of the function}

\item{sigma}{The standard deviation of the function}

\item{a}{The max x-coordinate of the area}
}
\value{
Area to terminal and graph of area under the curve
}
\description{
Area of Curve
}
\examples{
myncurve(mu=10, sigma=5, a=6)
}
