% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Number of Tickets}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{N number of seats available}

\item{gamma}{probability of overbooking}

\item{p}{probability that a passenger will show}
}
\value{
2 plots (discrete and continuous) representing the number of tickets
and a list of variable values

nd: The calculated number of tickets to sell using the discrete probability distribution.
nc: The calculated number of tickets to sell using the normal approximation.
N: The total number of seats available on the flight.
gamma: The maximum acceptable probability that the flight is overbooked.
p: The likelihood that a ticket holder will attend the flight.


The function also produces two plots showing the Objective function versus n. The Objective function
is derived by setting the defining equation to zero (e.g., 1 - gamma - pnorm(...) = 0). One plot represents
the discrete case, while the other illustrates the continuous case using the normal approximation.
}
\description{
Number of Tickets
}
\examples{
ntickets(400,.02,.95)
}
